pyproject.toml
[project]
name = "metaculus-bot"
version = "0.1.0"
description = "AI Forecasting Bot for Metaculus competitions"
authors = [
    {name = "Jordan Berman"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.76.0",
    "openai>=2.15.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "aiohttp>=3.13.3",
    "python-dotenv>=1.2.1",
    "litellm>=1.81.1",
    "aiosqlite>=0.22.1",
]

[project.scripts]
metaculus-bot = "main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "src"}]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
